# Assignment4/campuseatsOrders/src/main/java/com/example/campuseatsOrders/dto/OrderItemResponse.java
package com.example.campuseatsOrders.dto;

import java.math.BigDecimal;

public record OrderItemResponse(
        Long foodItemRef,
        String itemName,
        BigDecimal unitPrice,
        int quantity,
        BigDecimal lineTotal
) {
}